+++ src/public/app/services/ws.js
@@ -51,42 +51,6 @@ function logRows(syncRows) {
     if (filteredRows.length > 0) {
         console.debug(utils.now(), "Sync data: ", filteredRows);
     }
-
-    // Source - https://gist.github.com/jtsternberg/c272d7de5b967cec2d3d
-
-    var is_colliding = function( $div1, $div2 ) {
-        // Div 1 data
-        var d1_offset             = $div1.offset();
-        var d1_height             = $div1.outerHeight( true );
-        var d1_width              = $div1.outerWidth( true );
-        var d1_distance_from_top  = d1_offset.top + d1_height;
-        var d1_distance_from_left = d1_offset.left + d1_width;
-    
-        // Div 2 data
-        var d2_offset             = $div2.offset();
-        var d2_height             = $div2.outerHeight( true );
-        var d2_width              = $div2.outerWidth( true );
-        var d2_distance_from_top  = d2_offset.top + d2_height;
-        var d2_distance_from_left = d2_offset.left + d2_width;
-    
-        var not_colliding = ( d1_distance_from_top < d2_offset.top || d1_offset.top > d2_distance_from_top || d1_distance_from_left < d2_offset.left || d1_offset.left > d2_distance_from_left );
-    
-        // Return whether it IS colliding
-        return ! not_colliding;
-    };
-
-    let centerPane = document.getElementById("center-pane");
-    $("span").on("mouseenter", 
-        function(e) {
-            if (e.currentTarget.className === 'fancytree-title') {
-                if(is_colliding($(centerPane), $(e.currentTarget))) {
-                    e.currentTarget.title = e.currentTarget.innerText;
-                } else {
-                    e.currentTarget.title = "";
-                }
-            }
-        }
-    );
 }
 
 async function handleMessage(event) {
